Feat/transaction management

+++ Group7_iFINANCEAPP/Controllers/TransactionsController.cs
@@ -7,6 +7,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using Group7_iFINANCEAPP.Models;
+using Group7_iFINANCEAPP.Models.ViewModels;
 
 namespace Group7_iFINANCEAPP.Controllers
 {
@@ -17,9 +18,19 @@ namespace Group7_iFINANCEAPP.Controllers
         // GET: Transactions
         public ActionResult Index()
         {
-            var transaction = db.Transaction.Include(t => t.NonAdminUser);
-            return View(transaction.ToList());
+            var transactions = db.Transaction
+                                 .Include(t => t.TransactionLine)
+                                 .Include(t => t.NonAdminUser)
+                                 .ToList()
+                                 .Select(t => new TransactionWithLinesViewModel
+                                 {
+                                     Transaction = t,
+                                     TransactionLines = t.TransactionLine.ToList()
+                                 }).ToList();
+
+            return View(transactions);
         }
+
 
         // GET: Transactions/Details/5
         public ActionResult Details(int? id)
@@ -114,9 +125,20 @@ namespace Group7_iFINANCEAPP.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Transaction transaction = db.Transaction.Find(id);
-            db.Transaction.Remove(transaction);
-            db.SaveChanges();
+            Transaction transaction = db.Transaction
+                .Include(t => t.TransactionLine)
+                .SingleOrDefault(t => t.ID == id);
+
+            if (transaction != null)
+            {
+                foreach (var line in transaction.TransactionLine.ToList())
+                {
+                    db.TransactionLine.Remove(line);
+                }
+
+                db.Transaction.Remove(transaction);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
 
@@ -127,6 +149,23 @@ namespace Group7_iFINANCEAPP.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+
+        public ActionResult TransactionsWithLines()
+        {
+            using (var db = new Group7_iFINANCEDBEntities())
+            {
+                var data = db.Transaction
+                             .Include(t => t.TransactionLine)
+                             .ToList()
+                             .Select(t => new TransactionWithLinesViewModel
+                             {
+                                 Transaction = t,
+                                 TransactionLines = t.TransactionLine.ToList()
+                             }).ToList();
+
+                return View(data);
+            }
         }
     }
 }
